+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # ghc-tags-1.2 (2021-05-??)
-* Fix sorting of ctags
+* Fix sorting of ctags.
+* Express addresses of ctags as ex commands.
 
 # ghc-tags-1.1 (2021-05-18)
 * Fix compatibility with GHC 8.10.